backend/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
mongoengine==0.29.1
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
pymongo==3.11.4
python-decouple==3.8
pytz==2025.2
sqlparse==0.5.3
